# [SofaHelper] Speedup ReadAccessor conversion operator