--- a/SofaKernel/modules/SofaHelper/src/sofa/helper/accessor.h
+++ b/SofaKernel/modules/SofaHelper/src/sofa/helper/accessor.h
@@ -182,7 +182,7 @@ public:
     const_iterator end() const { return vref->end(); }
 
     ///////// Access the container for reading ////////////////
-    operator  const_container_type () const { return  *vref; }
+    operator const T& () const { return  *vref; }
     const_container_type* operator->() const { return vref; }
     const_container_type& operator* () const { return  *vref; }
     const_container_type& ref() const { return *vref; }          ///< this duplicate operator* (remove ?)