The implicit conversion operator returned a const value. Now, it's a const reference.

Before:


After:

```
------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                          Time             CPU   Iterations
------------------------------------------------------------------------------------------------------------------------------------
BM_Data_GetValue<int>                                                                           2.26 ns         2.26 ns    248888889
BM_Data_GetValue<sofa::type::vector<int>>/10000                                                 2.13 ns         2.15 ns    320000000
BM_Data_GetValue<sofa::type::vector<int>>/100000                                                2.14 ns         2.14 ns    298666667
BM_Data_GetValue<sofa::type::vector<int>>/1000000                                               2.14 ns         2.15 ns    320000000
BM_Data_GetValue<sofa::type::vector<sofa::type::Vec3>>/10000                                    2.00 ns         2.01 ns    373333333
BM_Data_GetValue<sofa::type::vector<sofa::type::Vec3>>/100000                                   1.94 ns         1.90 ns    344615385
BM_Data_GetValue<sofa::type::vector<sofa::type::Vec3>>/1000000                                  2.01 ns         2.01 ns    373333333
BM_Data_ReadAccessor_ImplicitConversion<int>                                                    2.88 ns         2.85 ns    235789474
BM_Data_ReadAccessor_ImplicitConversion<sofa::type::vector<int>>/10000                          2.81 ns         2.83 ns    248888889
BM_Data_ReadAccessor_ImplicitConversion<sofa::type::vector<int>>/100000                         3.02 ns         3.01 ns    248888889
BM_Data_ReadAccessor_ImplicitConversion<sofa::type::vector<int>>/1000000                        2.88 ns         2.83 ns    248888889
BM_Data_ReadAccessor_ImplicitConversion<sofa::type::vector<sofa::type::Vec3>>/10000             2.62 ns         2.61 ns    263529412
BM_Data_ReadAccessor_ImplicitConversion<sofa::type::vector<sofa::type::Vec3>>/100000            2.61 ns         2.61 ns    263529412
BM_Data_ReadAccessor_ImplicitConversion<sofa::type::vector<sofa::type::Vec3>>/1000000           2.58 ns         2.57 ns    280000000
BM_Data_ReadAccessor_ExplicitConversion<int>                                                    2.57 ns         2.55 ns    263529412
BM_Data_ReadAccessor_ExplicitConversion<sofa::type::vector<int>>/10000                          2.59 ns         2.55 ns    263529412
BM_Data_ReadAccessor_ExplicitConversion<sofa::type::vector<int>>/100000                         2.60 ns         2.61 ns    263529412
BM_Data_ReadAccessor_ExplicitConversion<sofa::type::vector<int>>/1000000                        2.64 ns         2.68 ns    280000000
BM_Data_ReadAccessor_ExplicitConversion<sofa::type::vector<sofa::type::Vec3>>/10000             2.81 ns         2.76 ns    248888889
BM_Data_ReadAccessor_ExplicitConversion<sofa::type::vector<sofa::type::Vec3>>/100000            2.85 ns         2.83 ns    248888889
BM_Data_ReadAccessor_ExplicitConversion<sofa::type::vector<sofa::type::Vec3>>/1000000           2.86 ns         2.83 ns    248888889
BM_Data_ReadAccessor_StarOperatorConversion<int>                                                2.66 ns         2.67 ns    263529412
BM_Data_ReadAccessor_StarOperatorConversion<sofa::type::vector<int>>/10000                      2.64 ns         2.67 ns    263529412
BM_Data_ReadAccessor_StarOperatorConversion<sofa::type::vector<int>>/100000                     2.59 ns         2.55 ns    263529412
BM_Data_ReadAccessor_StarOperatorConversion<sofa::type::vector<int>>/1000000                    2.57 ns         2.57 ns    280000000
BM_Data_ReadAccessor_StarOperatorConversion<sofa::type::vector<sofa::type::Vec3>>/10000         2.80 ns         2.83 ns    248888889
BM_Data_ReadAccessor_StarOperatorConversion<sofa::type::vector<sofa::type::Vec3>>/100000        3.03 ns         2.95 ns    248888889
BM_Data_ReadAccessor_StarOperatorConversion<sofa::type::vector<sofa::type::Vec3>>/1000000       2.81 ns         2.78 ns    235789474
```



______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).